nothing

+++ java/src/cs224n/features/DistanceBasedFeaturizer.java
@@ -51,7 +51,7 @@ public class DistanceBasedFeaturizer implements FeaturizerOne {
 					}
 				}
 			}
-			features.add(new FeatureValue(FEATURE_NAME, min/(passageTokenStrings.size()-1)));
+			features.add(new FeatureValue(FEATURE_NAME, (double) min/(passageTokenStrings.size()-1)));
 			return features;
 		}
 	}
